+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-## v3.0.1 [2019-12-11]
+## In Progress
 
 __Breaking Changes:__